pyproject.toml
[project]
name = "oregon-python-sso"
version = "0.1.0"
description = "sso"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic~=1.15.0",
    "asyncpg~=0.31.0",
    "fastapi~=0.135.1",
    "httpx~=0.28.1",
    "opentelemetry-api>=1.41.1",
    "opentelemetry-exporter-otlp>=1.41.1",
    "opentelemetry-instrumentation>=0.62b1",
    "opentelemetry-instrumentation-fastapi>=0.62b1",
    "opentelemetry-sdk>=1.41.1",
    "passlib[bcrypt]~=1.7.4",
    "pydantic[email]~=2.12.5",
    "pyjwt~=2.12.1",
    "python-dotenv~=1.1.0",
    "sqlalchemy~=2.0.48",
    "structlog>=25.5.0",
    "uvicorn[standard]~=0.34.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
    "ty>=0.0.21",
]

[tool.ruff]
target-version = "py314"
line-length = 100
include = ["src/**/*.py", "tests/**/*.py"]
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.pyi",
    "migrations",
    "node_modules",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
    "RET",
    "ARG",
    "PTH",
    "RUF",
    "ANN",
    "ERA",
    "G",
]

ignore = [
    "E501",
    "B008",
]

fixable = ["ALL"]

[tool.ty.src]
include = ["src", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
asyncio_mode = "auto"
